refactor(statusline): rename file
